etl/requirements-wmaa.txt
apache-airflow[postgres]
psycopg2
